% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactions.R
\name{extractInteractions}
\alias{extractInteractions}
\title{Extract Interactions}
\usage{
extractInteractions(object, ...)
}
\arguments{
\item{object}{An object of class \code{"pmoe"}.}

\item{...}{Currently unused.}
}
\description{
Extract interactions.
}

